requirements.txt
asgiref==3.3.1
astroid==2.5.1
attrs==20.3.0
certifi==2020.12.5
chardet==4.0.0
Django==3.1.7
djangorestframework==3.12.2
flake8==3.9.0
idna==2.10
iniconfig==1.1.1
isort==5.7.0
lazy-object-proxy==1.5.2
mccabe==0.6.1
packaging==20.9
pluggy==0.13.1
py==1.10.0
pycodestyle==2.7.0
pyflakes==2.3.0
pylint==2.7.2
pylint-django==2.4.2
pylint-plugin-utils==0.6
pyparsing==2.4.7
pytest==6.2.2
pytest-django==4.1.0
pytz==2021.1
requests==2.25.1
sqlparse==0.4.1
toml==0.10.2
urllib3==1.26.3
wrapt==1.12.1
